requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
googlesearch_python==1.2.3
numpy==1.26.4
pandas==2.2.2
python_dateutil==2.8.2
Requests==2.31.0
scikit_learn==1.3.1
urllib3==2.2.1
whois==0.9.27
gunicorn==20.1.0
google-generativeai==0.3.2
